LicenseInfoResolver: Move toResolvedCopyrights() to ResolvedLicenseInfo
ResolvedLicenseInfo has become the only caller. Thus, make it also
private.
