+++ model/src/main/kotlin/licenses/ResolvedLicenseInfo.kt
@@ -238,3 +238,21 @@ fun List<ResolvedLicense>.filterExcluded() =
         resolvedLicense.sources != setOf(LicenseSource.DETECTED) ||
                 resolvedLicense.locations.any { it.matchingPathExcludes.isEmpty() }
     }.map { it.filterExcludedCopyrights() }
+
+private fun Collection<ResolvedCopyrightFinding>.toResolvedCopyrights(process: Boolean): List<ResolvedCopyright> {
+    val allStatements = map { it.statement }
+    val processedStatements = if (process) {
+        CopyrightStatementsProcessor().process(allStatements).toMap()
+    } else {
+        allStatements.associateBy({ it }, { mutableSetOf(it) })
+    }
+
+    return processedStatements.mapValues { (_, originalStatements) ->
+        filter { it.statement in originalStatements }
+    }.filterValues { it.isNotEmpty() }.entries.map() { (statement, findings) ->
+        ResolvedCopyright(statement, findings.toSet())
+    }
+}
+
+private fun CopyrightStatementsProcessor.Result.toMap(): Map<String, Set<String>> =
+    processedStatements + unprocessedStatements.associateWith { setOf(it) }
